chatbots/rag_bot/requirements.txt
flask
python-dotenv
openai
faiss-cpu
PyPDF2
numpy